+++ db/migrate/20260510130100_enable_rls_on_therapist_messages.rb
@@ -1,0 +1,9 @@
+class EnableRlsOnTherapistMessages < ActiveRecord::Migration[8.1]
+  def up
+    execute "ALTER TABLE #{quote_table_name(:therapist_messages)} ENABLE ROW LEVEL SECURITY"
+  end
+
+  def down
+    execute "ALTER TABLE #{quote_table_name(:therapist_messages)} DISABLE ROW LEVEL SECURITY"
+  end
+end